Testsuite: add a --no-auto-path option
In the production environment, we want the testsuite to exercize the
adareducer that is being packaged (installed in the environment), not
the adareducer that is part of the testsuite source tree.

This new option prevents the testsuite from automatically adding the
adareducer part of the testsuite source tree to the environment.

TN: V317-010